fix: refactor similarproduct component
