+++ components/product/product-details/SimilarProducts/constants.ts
@@ -1,0 +1,16 @@
+// constants.ts
+export const PROPERTY_ORDER = [
+  "Altura da Lente",
+  "Largura da Lente",
+  "Largura da Ponte",
+  "Hastes",
+  "Frente Total",
+  "Aro",
+] as const;
+
+export const PROPERTY_NAME_MAP = {
+  "Altura da Lente": "Altura",
+  "Largura da Lente": "Largura",
+  "Largura da Ponte": "Ponte",
+  "Frente Total": "Frente",
+} as const;
