+++ fresh.gen.ts
@@ -12,6 +12,7 @@ import * as $Carrossel from "./islands/Carrossel.tsx";
 import * as $CartModalMobile from "./islands/CartModalMobile.tsx";
 import * as $CategoryText from "./islands/CategoryText.tsx";
 import * as $ChooseLensButton from "./islands/ChooseLensButton.tsx";
+import * as $ColorSelectorSimilar from "./islands/ColorSelectorSimilar.tsx";
 import * as $ContactForm from "./islands/ContactForm.tsx";
 import * as $CouponListScript from "./islands/CouponListScript.tsx";
 import * as $FaqContainer from "./islands/FaqContainer.tsx";
@@ -58,6 +59,7 @@ const manifest = {
     "./islands/CartModalMobile.tsx": $CartModalMobile,
     "./islands/CategoryText.tsx": $CategoryText,
     "./islands/ChooseLensButton.tsx": $ChooseLensButton,
+    "./islands/ColorSelectorSimilar.tsx": $ColorSelectorSimilar,
     "./islands/ContactForm.tsx": $ContactForm,
     "./islands/CouponListScript.tsx": $CouponListScript,
     "./islands/FaqContainer.tsx": $FaqContainer,
