+++ islands/ColorSelectorSimilar.tsx
@@ -1,0 +1,3 @@
+export {
+    default,
+} from "$store/components/product/product-details/SimilarProducts/ColorSelector.tsx";
