+++ islands/SimilarProducts.tsx
@@ -1,1 +1,4 @@
-export { default, loader } from "$store/components/product/product-details/SimilarProducts.tsx";
+export {
+    default,
+    loader,
+} from "$store/components/product/product-details/SimilarProducts/index.tsx";
