+++ static/tailwind.css
@@ -8637,8 +8637,8 @@ details.collapse summary::-webkit-details-marker {
     margin-bottom: 0.25rem;
   }
 
-  .lg\:mb-10 {
-    margin-bottom: 2.5rem;
+  .lg\:mb-12 {
+    margin-bottom: 3rem;
   }
 
   .lg\:mb-16 {
